refactor: merge event parser families back together
